worknig sensor

+++ db/migrate/20100407225942_create_sensors.rb
@@ -2,7 +2,7 @@ class CreateSensors < ActiveRecord::Migration
   def self.up
     create_table :sensors do |t|
       t.string :name
-      t.double :temp
+      t.float :temp
 
       t.timestamps
     end
